+++ docs/conf.py
@@ -20,7 +20,7 @@
 from datetime import datetime
 import re
 
-_version = '2.8.1'
+_version = '2.9.0'
 
 
 def isBoostFunc(what, obj):