+++ docs/changelog.rst
@@ -1,6 +1,16 @@
 Changelog
 =========
 
+Unreleased
+----------
+
+* Add the option to only show progress on the dashboard. (`#107`_)
+* Import escape directly from markupsafe, instead of from flask. (`#106`_)
+
+.. _#108: https://github.com/sybrenjansen/mpire/pull/107
+.. _#107: https://github.com/sybrenjansen/mpire/issues/106
+
+
 2.8.1
 -----