+++ docs/changelog.rst
@@ -11,7 +11,7 @@ Unreleased
 * When using insights on Windows the arguments of the top 5 longest tasks are now available as well.
 * Fixed deprecated ``escape`` import from ``flask`` by importing directly from ``markupsafe``. (`#106`_)
 
-.. _#98: https://github.com/sybrenjansen/mpire/issues/96
+.. _#96: https://github.com/sybrenjansen/mpire/issues/96
 .. _#107: https://github.com/sybrenjansen/mpire/pull/107
 .. _#104: https://github.com/sybrenjansen/mpire/issues/104
 .. _#106: https://github.com/sybrenjansen/mpire/issues/106